+++ jeap-process-archive-adapter-objectstorage/src/test/java/ch/admin/bit/jeap/processarchive/objectstorage/CustomObjectStorageStrategyIT.java
@@ -1,5 +1,6 @@
 package ch.admin.bit.jeap.processarchive.objectstorage;
 
+import ch.admin.bit.jeap.processarchive.domain.archive.ArchiveDataObjectStoreStorageException;
 import ch.admin.bit.jeap.processarchive.domain.archive.ArchiveDataStorageInfo;
 import ch.admin.bit.jeap.processarchive.domain.archive.lifecycle.LifecyclePolicyService;
 import ch.admin.bit.jeap.processarchive.plugin.api.archivedata.ArchiveData;
@@ -7,6 +8,7 @@ import ch.admin.bit.jeap.processarchive.plugin.api.archivedata.schema.ArchiveDat
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
@@ -16,6 +18,8 @@ import java.util.UUID;
 import static ch.admin.bit.jeap.processarchive.objectstorage.ObjectStorageConfiguration.JEAP_PAS_TEST_INMEMORY_PROFILE;
 import static java.util.Collections.emptyList;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 @ActiveProfiles(JEAP_PAS_TEST_INMEMORY_PROFILE)
 @SpringBootTest(classes = {CustomObjectStorageStrategyConfig.class, ObjectStorageConfiguration.class, HashProviderTestConfig.class},
@@ -25,11 +29,23 @@ class CustomObjectStorageStrategyIT {
     @Autowired
     ArchiveDataObjectStoreAdapter archiveDataObjectStoreAdapter;
 
+    @Autowired
+    ObjectStorageProperties objectStorageProperties;
+
     @MockitoBean
     LifecyclePolicyService lifecyclePolicyService;
 
     private final static ArchiveDataSchema ARCHIVE_DATA_SCHEMA = ArchiveDataSchema.builder()
             .schemaDefinition("test".getBytes(StandardCharsets.UTF_8))
+            .system("test-system")
+            .name("schemaname")
+            .referenceIdType("ch.admin.bit.jeap.audit.type.SchemaNameArchive")
+            .version(1)
+            .fileExtension("avpr")
+            .build();
+
+    private final static ArchiveDataSchema OTHER_ARCHIVE_DATA_SCHEMA = ArchiveDataSchema.builder()
+            .schemaDefinition("different-schema-test".getBytes(StandardCharsets.UTF_8))
             .system("test-system")
             .name("schemaname")
             .referenceIdType("ch.admin.bit.jeap.audit.type.SchemaNameArchive")
@@ -47,6 +63,39 @@ class CustomObjectStorageStrategyIT {
         assertThat(storageInfo.getKey()).startsWith(CustomObjectStorageStrategy.CUSTOM_PREFIX);
     }
 
+    @Test
+    void testStoreTwice_sameSchema_shouldNotThrowException() {
+        ArchiveData archiveData = createArchiveData();
+
+        archiveDataObjectStoreAdapter.store(archiveData, ARCHIVE_DATA_SCHEMA);
+        assertDoesNotThrow(() ->
+                archiveDataObjectStoreAdapter.store(archiveData, ARCHIVE_DATA_SCHEMA));
+    }
+
+    @Test
+    void testStoreTwice_differentSchema_overwriteAllowed_shouldNotThrowException() {
+        ArchiveData archiveData = createArchiveData();
+
+        archiveDataObjectStoreAdapter.store(archiveData, ARCHIVE_DATA_SCHEMA);
+        assertDoesNotThrow(() ->
+                archiveDataObjectStoreAdapter.store(archiveData, OTHER_ARCHIVE_DATA_SCHEMA));
+    }
+
+    @DirtiesContext(methodMode = DirtiesContext.MethodMode.AFTER_METHOD)
+    @Test
+    void testStoreTwice_differentSchema_overwriteNotAllowed_shouldThrowException() {
+        objectStorageProperties.setSchemaOverwriteAllowed(false);
+
+        ArchiveData archiveData = createArchiveData();
+
+        archiveDataObjectStoreAdapter.store(archiveData, ARCHIVE_DATA_SCHEMA);
+        assertThatThrownBy(() ->
+                archiveDataObjectStoreAdapter.store(archiveData, OTHER_ARCHIVE_DATA_SCHEMA))
+                .isInstanceOf(ArchiveDataObjectStoreStorageException.class)
+                .hasCauseInstanceOf(S3ObjectStorageException.class)
+                .hasMessageContaining("Schema overwrite not allowed");
+    }
+
     private ArchiveData createArchiveData() {
         return ArchiveData.builder()
                 .contentType("application/json")
@@ -58,5 +107,4 @@ class CustomObjectStorageStrategyIT {
                 .metadata(emptyList())
                 .build();
     }
-
 }